+++ modules/config.py
@@ -1,0 +1,8 @@
+#Def Vars
+isDoing = True
+explode = False
+module = "EMPTY"
+
+serialNum = "EMPTY"
+batteries = "EMPTY"
+litIndicators = ["EMPTY"]
